addressing comments to pull request #21514 for GH issue #18644
